requirements.txt
flask
mongoengine
flask-mongoengine
python-dotenv
langchain-ollama
langchain
chromadb
ultralytics
flask-login
flask-wtf
passlib==1.7.4
bcrypt==4.1.2
Werkzeug==3.0.1
xhtml2pdf>=0.2.9
scikit-learn
pandas
numpy